move guard under guard directory
