+++ src/auth/guard/auth.guard.ts
@@ -4,10 +4,10 @@ import {
   ExecutionContext,
   UnauthorizedException,
 } from '@nestjs/common';
-import { FirebaseService } from './firebase/firebase.service';
-import { PrismaService } from '../common/services/prisma.service';
-import { UserInfo } from './type/user-info.type';
-import { RequestWithAuth } from './type/request-with-auth';
+import { FirebaseService } from '../firebase/firebase.service';
+import { PrismaService } from '../../common/services/prisma.service';
+import { UserInfo } from '../type/user-info.type';
+import { RequestWithAuth } from '../type/request-with-auth';
 
 @Injectable()
 export class FirebaseAuthGuard implements CanActivate {
